Tweak lint rules
Repo's getting popular. From now on, nothing but good commit messages!

Add mocha env, so that `describe` and `it` don't trigger warning about
globals.

Remove prefer-const. `let a = {}; a.b = 1;` previously warned about
switching to `const`. But personally I want const to indicate nothing
but "this is totally immutable". This'll be the convention of this repo.
Use `let` even if you mutate a nested collection.