Correctly fix raf-ing after being unmounted
Attempted fix: b2e617e

That one isn't a good fix, because the check of
`this.unsubscribeAnimation != null` in `animationRender` also prevented
the `setState` from being done one last time for the last value (since
`unsubscribeAnimation` is already set to `null` in `animationStep`,
after which we return the (never rendered) final state).

The solution is to put a temporary boolean flag until we find better
ways ot solve this. This is the equivalent of `this.isMounted()` React
API, except the latter will soon be deprecated.