Fix TransitionSpring not entering new keys
The good thing we did is to set the merged `endValue[key]` to the value
of `willEnter(key)` for the first tick. This ensured that we rendered the
initial value once. However, this caused `newCurrVelocity` to be `0` on
that tick (stepping from x to x so no velocity). This would unsubscribe
the animation. The (temporary) solution is to check whether we added a
new key with a flag. If not (and if velocity is `0`), we unsubscribe for
real.