Use correct terminology for endValue cb
Previously we call the value passed to `endValue` cb as
`currentPosition`. This is misleading since it's really the value
returned by previous call to `endValue`. We correct the term here.

Also, we pass the object to `endValue` as much as we can (every case
where we don't really use `prevValue` argument). Passing obj will
potentially be optimized by Spring in the future (cached) vs uncacheable
for an `endValue` function callback (since it can change at each call,
  through using `prevValue`).