Set default time interval to 1000/60 instead of 16.
This will prevent the animationLoop to get "out of sync". Because the
accumulated time, and the decision to "flush" a frame is based off
whether the accumulated time is > 16.66666ms and we used to increment
"time" by only 16ms, things would get a little out of sync. This would
break the consistency of the values in the tests.