Check in new Motion component
Last part of the 3 commits.
This is somewhat of a rewrite that fixes a bunch of bugs and confusion
points (e.g. #179, #88, #90). The whole thing's typed =D.

The animation loop's logic was preserved but rewritten. Right now it
resides inside Motion directly, but once we finish refactoring
StaggerMotion and TransitionMotion, we'll abstract it out again if
there's any common part.

The motivation was that the old separation of animation loop and
components was visual and not really conceptual. We still end up passing
stuff back and forth between the components and the animation loop.
Furthermore, `requestAnimationFrame` itself is already basically an
animation loop. Our didn't have much benefit, but I could be wrong.
Anyhow, the new Motion is hack-free and cleaner.

The tests catch a bunch of edge-cases now too. The numbers mostly
correspond to the previous animation loop's, with the exception of a few
js float rounding errors (due to tweaks to new logic).

There are a few low-hanging fruits in terms of perf improvement. We'll
work on these once StaggerMotion and TransitionMotion have been
converted too.