(Continued) TransitionMotion tests Array styles
Part 3 of the changes, check previous commit for the component's logic
changes.

This part is big but mostly brain-dead conversion from object to array.

It fixes a test bug that previous leaked through for the test "should behave
well when many owner updates come in-between rAFs" and the next one.
Previously, `setState({a: ...})` didn't remove the key b (obviously).
The intention was to remove it.